Add cache based on request hash
Alter the download_cds command to download files to a user specified
cache directory where files are named after
`get_json_sem_hash(request)`. The `RequestEntry` class is introduce to
keep the data necessary for producing the hash collected.

Output according to `filename_spec` are created using symlinks to the
cache.

Main motivation for this is to allow multiple projects that might
have overlapping data to re-use previously downloaded requests.